# swagger.yaml
openapi: 3.0.3
info:
  title: Photo Studio Booking - API
  version: 1.1.0
  description: >
    API Gateway (5000) for booking & payments + integration endpoints.
    Payment webhook is served by Payment Service (5002).
    Message Broker (5050) и Notification Service (5004) описаны как отдельные servers.
servers:
  - url: http://localhost:5000
    description: API Gateway
  - url: http://localhost:5002
    description: Payment Service (webhooks)
  - url: http://localhost:5050
    description: Message Broker
  - url: http://localhost:5004
    description: Notification Service
paths:
...
  /api/bookings/availability:
    get:
      summary: Check availability slots
      tags: [Bookings]
      parameters:
        - in: query
          name: hall_id
          required: false
          schema: { type: string, example: hall-001 }
        - in: query
          name: start_date
          required: true
          schema: { type: string, format: date-time, example: "2025-12-17T00:00:00Z" }
        - in: query
          name: end_date
          required: true
          schema: { type: string, format: date-time, example: "2025-12-18T00:00:00Z" }
      responses:
        "200":
          description: Slots list
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items: { $ref: "#/components/schemas/Slot" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/bookings:
    post:
      summary: Create booking
      tags: [Bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingCreateRequest" }
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400":
          description: Selected time not available
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/bookings/{booking_id}:
    get:
      summary: Get booking by id
      tags: [Bookings]
      parameters:
        - in: path
          name: booking_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      summary: Cancel booking
      tags: [Bookings]
      parameters:
        - in: path
          name: booking_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "400":
          description: Already cancelled / invalid state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/payments:
    post:
      summary: Create payment
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentCreateRequest" }
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/payments/{payment_id}:
    get:
      summary: Get payment by id
      tags: [Payments]
      parameters:
        - in: path
          name: payment_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/payments/{payment_id}/refund:
    post:
      summary: Refund payment (only succeeded)
      description: >
        Before calling this endpoint, you must create a payment and process
        successful webhook so that payment status becomes `succeeded`.
      tags: [Payments]
      parameters:
        - in: path
          name: payment_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundRequest"
      responses:
        "200":
          description: Refund result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Refund" }
        "400":
          description: Invalid state or validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Выбранное время недоступно"
    Slot:
      type: object
      properties:
        hall_id: { type: string, example: hall-001 }
        start_time: { type: string, example: "Wed, 17 Dec 2025 23:00:00 GMT" }
        end_time: { type: string, example: "Wed, 17 Dec 2025 23:15:00 GMT" }
        available: { type: boolean, example: true }
    BookingCreateRequest:
      type: object
      required: [hall_id, user_id, start_time, end_time, customer_name, customer_email, customer_phone]
      properties:
        hall_id: { type: string, example: hall-001 }
        user_id: { type: string, example: user-123 }
        start_time: { type: string, format: date-time, example: "2025-12-17T14:00:00" }
        end_time: { type: string, format: date-time, example: "2025-12-17T16:00:00" }
        customer_name: { type: string, example: "Иван Иванов" }
        customer_email: { type: string, example: "ivan@example.com" }
        customer_phone: { type: string, example: "+79161234567" }
    Booking:
      type: object
      properties:
        booking_id: { type: string, format: uuid }
        hall_id: { type: string, example: hall-001 }
        user_id: { type: string, example: user-123 }
        start_time: { type: string, example: "Wed, 17 Dec 2025 14:00:00 GMT" }
        end_time: { type: string, example: "Wed, 17 Dec 2025 16:00:00 GMT" }
        status:
          type: string
          enum: [pending_payment, confirmed, cancelled, completed]
          example: pending_payment
        total_amount: { type: string, example: "3000.0" }
        customer_name: { type: string, example: "Иван Иванов" }
        customer_email: { type: string, example: "ivan@example.com" }
        customer_phone: { type: string, example: "+79161234567" }
        created_at: { type: string, example: "Tue, 16 Dec 2025 23:37:11 GMT" }
        updated_at: { type: string, nullable: true, example: null }
    PaymentCreateRequest:
      type: object
      required: [booking_id, amount, payment_type, payment_method]
      properties:
        booking_id: { type: string, format: uuid }
        amount: { type: number, example: 3000.0 }
        payment_type:
          type: string
          enum: [prepayment, full_payment]
          example: full_payment
        payment_method:
          type: string
          enum: [yookassa, sberpay, tinkoff]
          example: yookassa
        return_url:
          type: string
          nullable: true
          example: "https://example.com/return"
    Payment:
      type: object
      properties:
        payment_id: { type: string, format: uuid }
        booking_id: { type: string, format: uuid }
        amount: { type: string, example: "3000.0" }
        status:
          type: string
          enum: [pending, succeeded, failed, refunded]
          example: pending
        payment_method:
          type: string
          enum: [yookassa, sberpay, tinkoff]
        payment_url: { type: string, example: "https://yoomoney.ru/checkout/payments/v2/contract?orderId=..." }
        external_payment_id: { type: string, example: "yookassa-0294d913-3de5-4ad0-8616-4eb6b7b233a1" }
        created_at: { type: string, example: "Tue, 16 Dec 2025 23:39:28 GMT" }
        updated_at: { type: string, nullable: true, example: null }
    RefundRequest:
      type: object
      required: [payment_id]
      properties:
        payment_id:
          type: string
          format: uuid
          example: "af50a1ae-494a-465c-80af-446406a935ba"
        amount:
          type: number
          nullable: true
          example: 1500.0
        reason:
          type: string
          nullable: true
          example: "Partial refund"
    Refund:
      type: object
      properties:
        refund_id: { type: string, format: uuid }
        payment_id: { type: string, format: uuid }
        amount: { type: string, example: "1500.0" }
        status: { type: string, example: succeeded }
        timestamp: { type: string, format: date-time }
...